TaskHandler Implemented
Lumberjack now arrives at forest region when assigned lumberjack and cuts all trees down within that region.

Next I need to create logic so that the tree spawns wood resources and has a chance to spawn tree saplings. I also need to create a task to collect resources once a tree is cut down. The lumberjack needs to collect the wood resource and take it to the nearest storage building (right now that is the house that is spawned)